rewrite heartrate process function and devide function.c into several files according to their functions
