+++ ibaby_smarthome_multinode/src/wearable_node/function/lwm2m.h
@@ -1,0 +1,7 @@
+#ifndef LWM2M_H
+#define LWM2M_H
+
+/** function for starting lwm2mClient */
+extern int   lwm2m_client_start(void);
+
+#endif
