+++ ibaby_smarthome_multinode/src/wearable_node/function/process_acc.c
@@ -56,203 +56,70 @@
  * \addtogroup	EMBARC_APP_FREERTOS_IBABY_SMARTHOME_NODES_WEARABLE_NODE
  * @{
  */
+/* standard C HAL */
+#include <math.h>
+
+/* embARC HAL */
+#include "embARC.h"
+#include "embARC_debug.h"
+
 /* custom HAL */
-#include "function.h"
+#include "common.h"
+#include "process_acc.h"
+#include "print_msg.h"
+#include "acceleration.h"
 
+/**
+ * \name	local macro for function of wearable node
+ * @{
+ */
+#define WARN_ACCL_Z    (-8) /*!< lower value of warning acceleration */
 
-/*
-**************************************************************
-*  This part will be deleted in release version
-*/
-#if USED_TIMER1
-/** arc timer 1 interrupt routine */
-static void timer1_isr(void *ptr)
-{
-	timer_int_clear(TIMER_1);
-	t1_count++;
-}
+/*!< parameters of low pass filter for latest acceleration value */
+#define PAR_ACC_BASE   (30) /*!< base value of low pass filter coefficient */
+#define PAR_ACC_STEP   (5)  /*!< step of coefficient enlarge */
+#define CNT_ACC_STEP   (5)  /*!< step of counter enlarge */
+#define THOLD_ACC_DIFF (300)/*!< threshold of acceleration value variation range, counter enlarge */
+#define THOLD_ACC_CNT  (30) /*!< threshold of counter, coefficient enlarge  */
 
-/** print message for debug major function */
-extern void timer1_start(void)
-{
-	t1_count = 0;
+/*!< parameters of low pass filter for latest motion intersity */
+#define PAR_SVM_BASE   (80) /*!< base value of low pass filter coefficient */
+#define PAR_SVM_STEP   (20) /*!< step of coefficient enlarge */
+#define CNT_SVM_STEP   (5)  /*!< step of counter enlarge */
+#define THOLD_SVM_DIFF (30) /*!< threshold of acceleration value variation range, counter enlarge */
+#define THOLD_SVM_CNT  (50) /*!< threshold of counter, coefficient enlarge */
 
-	int_handler_install(INTNO_TIMER1, timer1_isr);
-	int_enable(INTNO_TIMER1);
-	timer_start(TIMER_1, TIMER_CTRL_IE, T1_COUNT_UNIT_LMS);
-}
+/*!< parameters of limiting filter */
+#define MAX_LIMIT_SVM (200) /*!< upper limit of svm */
+#define MIN_LIMIT_SVM (5)   /*!< lower limit of svm */
 
-/** print message for debug major function */
-extern uint32_t timer1_stop(void)
-{
-	uint32_t dec;
+#define THOLD_INTEN_AW (1000)/*!< threshold of motion intensity */
+#define LEN_STA_WAKE   (5)   /*!< length of state queue indicated baby awake */
+#define LEN_STA_SLEEP  (LEN_STA_QUEUE - LEN_STA_WAKE)
+#define PAR_STA_AW     (0.75)/*!< proportion of wake-state in total queue indicated baby awake */
 
-	timer_stop(TIMER_1);
-	dec = t1_count - (t1_count/10) * 10;
+/*!< sparameters of Webster sleep-wake determine method */
+#define P  (0.0046) /*!< 0.0046:coefficient of scale */
+#define K0 (0.04)   /*!< coefficient of relationship */
+#define K1 (0.22)
+#define K2 (1)
+#define K3 (0.24)
+#define K4 (0.06)
 
-	// EMBARC_PRINTF("************     timing     ************\r\n");
-	// EMBARC_PRINTF("* timer1 counter : %d.%d ms\r\n", t1_count/10, dec);
-	// EMBARC_PRINTF("****************************************\r\n\r\n");
+/*!< variable of low pass filter for raw acceleration data */
+static char cnt_x, cnt_y, cnt_z;                /*!< low pass filter counter */
+static unsigned char par_x, par_y, par_z;       /*!< low pass filter coefficient */
+static bool flag_old_x, flag_old_y, flag_old_z; /*!< flag of last change direction of value */
 
-	return dec;
-}
-#endif/* USED_TIMER1 */
-/*
-*  end of this part
-**************************************************************
-*/
+/*!< variable of calculating intensity of motion in one sampling period */
+static int x_old, y_old, z_old; /*!< last value */
+static int svm_old;             /*!< SVM : signal vector magnitude for difference */
 
+/*!< variable of low pass filter for svm data */
+static char cnt_v;          /*!< low pass filter counter */
+static unsigned char par_v; /*!< low pass filter coefficient */
+static bool flag_old_v;     /*!< flag of change direction of value */
 
-#if PRINT_DEBUG_FUNC
-/** print message for primary function */
-extern void print_msg_func(void)
-{
-	char str[150];
-
-	EMBARC_PRINTF("\r\n************ Primary function ************\r\n");
-	sprintf(str,
-                "* Body heartrate   : %dbpm\r\n* Body temperature : %d.%d'C\r\n* Motion intensity : %d\r\n", 
-                data_report_wn.hrate,
-                data_report_wn.btemp/10, data_report_wn.btemp%10,
-                data_report_wn.motion_intensity);
-	EMBARC_PRINTF(str);
-	
-	if (data_report_wn.state == SLEEP) {
-		EMBARC_PRINTF("* State : sleep\r\n");
-	} else {
-		EMBARC_PRINTF("* State : wake\r\n");
-	}
-
-	if (data_report_wn.warn_btemp) {
-		EMBARC_PRINTF("* Warn of abnormal body temperature\r\n");
-	}
-
-	if (data_report_wn.warn_hrate) {
-		EMBARC_PRINTF("* Warn of abnormal body heartrate\r\n");
-	}
-
-	if (data_report_wn.warn_downward) {
-		EMBARC_PRINTF("* Warn of sleep on his stomach\r\n");
-	}
-
-	if (data_report_wn.event_awake == AWAKE) {
-		EMBARC_PRINTF("* Baby awake!\r\n");
-	}
-
-	EMBARC_PRINTF("\r\n");
-}
-#endif /* PRINT_DEBUG_FUNC */
-
-#if PRINT_DEBUG_AWAKE
-/** print message for debug awake event detecting function */
-extern void print_msg_awake(void)
-{
-	char str[50];
-
-	EMBARC_PRINTF("\r\n************ Awake detecting ************\r\n");
-	sprintf(str, "* Motion intensity in 5s : %d\r\n", sum_svm_5s);
-	EMBARC_PRINTF(str);
-
-	for (uint i = 0; i < LEN_STA_QUEUE; ++i) {
-		if (i != 0 && !(i%3)) {
-			EMBARC_PRINTF("\r\n");
-		}
-		
-		if (state_aw[i]) {
-			sprintf(str, "* state %d : sleep\t", i);
-			EMBARC_PRINTF(str);
-		} else {
-			sprintf(str, "* state %d : wake\t", i);
-			EMBARC_PRINTF(str);
-		}
-	}
-	EMBARC_PRINTF("\r\n");
-
-	if (data_report_wn.event_awake == AWAKE) {
-		EMBARC_PRINTF("* Baby awake!\r\n\r\n");
-	}
-}
-#endif /* PRINT_DEBUG_AWAKE */
-
-#if PRINT_DEBUG_SLEEP
-/** print message for debug Sleep-Wake state monitoring function */
-extern void print_msg_sleep(uint state)
-{
-	char str[50];
-
-	EMBARC_PRINTF("\r\n************ Sleep monitoring ************\r\n");
-	for (uint i = 0; i < 5; ++i) {
-		if (i == 2 || i == 3) {
-			EMBARC_PRINTF("\r\n");
-		}
-		sprintf(str, "* Intensity %d : %d\t\t", i, inten_sl[i]/100);
-		EMBARC_PRINTF(str);
-	}
-
-	sprintf(str, "\r\n* Intensity score in 5min : %f\r\n", score_sl);
-	EMBARC_PRINTF(str);
-
-	if (!state) {
-		EMBARC_PRINTF("* State of 2min ago : wake\r\n\r\n");
-	} else {
-		EMBARC_PRINTF("* State of 2min ago : sleep\r\n\r\n");
-	}
-}
-#endif /* PRINT_DEBUG_SLEEP */
-
-/* function for deal with heartrate by filter */
-extern void process_hrate(uint32_t *hrate)
-{
-	static int rates[HRATE_SIZE]; /* array of heart rates */
-	static int rate_spot = 0;
-	static int last_beat = 0;     /* Time at which the last beat occurred */
-
-	static float beats_per_min;
-	static int beat_aver;
-
-	static bool flag_timer_stop = true;
-
-	int time_cur = 10, ir_value = 0, delta = 1;
-	uint32_t data_rdy;
-
-	data_rdy = hrate_sensor_read(&ir_value);
-
-	if (data_rdy == E_OK && check_beat(ir_value) == 1)
-	{
-		flag_timer_stop != flag_timer_stop;
-		// printf("%d\n", ir_value);
-
-		// time_cur = this is a function that get system current time
-		// delta = time_cur - last_beat;
-		last_beat = time_cur;
-
-		if (flag_timer_stop == true)
-		{
-			delta = 2 * t1_count;
-			t1_count = 1;
-			beats_per_min = 60 / (delta / 10000.0);
-
-			// printf("%f\n", beats_per_min);
-
-			if (beats_per_min < 255 && beats_per_min > 20)
-			{
-				rates[rate_spot++] = (int)beats_per_min; /* store this reading in the array */
-				rate_spot %= HRATE_SIZE;                 /* wrap variable */
-
-				//take average of readings
-				beat_aver = 0;
-				for (int x = 0 ; x < HRATE_SIZE ; x++)
-					beat_aver += rates[x];
-				beat_aver /= HRATE_SIZE;
-			}
-			printf("%d\n", beat_aver);
-		} else {
-			t1_count = 1;
-		}
-	}
-
-	*hrate = (uint32_t)beat_aver;
-}
 
 /* function for deal with acclerate by filter */
 static int filter_acc(int val_new,
@@ -554,83 +421,6 @@ extern bool func_detect_downward(float acc_temp)
 	return warn;
 }
 
-#if LWM2M_CLIENT
-/** task for lwm2m client */
-static void task_lwm2m_client(void *par)
-{
-	unsigned int cpu_status;
-	
-	while (1) {
-		lwm2m_client_conn_stat = 1;
 
-		/* get into sub function(lwm2mclient) to do all work about data interaction with gateway(lwm2m server) */
-		if (lwm2mclient(&c_info) == 0) {
-			EMBARC_PRINTF("LwM2M client end successfully\r\n");
-		} else {
-			EMBARC_PRINTF("Error: LwM2M client end failed\r\n");
-		}
-
-		cpu_status = cpu_lock_save();
-		if (c_info.server != NULL)     free((void *)(c_info.server));
-		if (c_info.ep_name != NULL)    free((void *)(c_info.ep_name));
-		if (c_info.serverPort != NULL) free((void *)(c_info.serverPort));
-		lwm2m_client_conn_stat = 0;
-		cpu_unlock_restore(cpu_status);
-
-		vTaskSuspend(task_lwm2m_client_handle);
-	}
-}
-/** function for initialize and start lwm2m client */
-extern int lwm2m_client_start(void)
-{
-	int c_quit = 0;
-	task_lwm2m_client_handle = 0;
-	
-	/* check to see if wifi works */
-	if (!lwip_pmwifi_isup()) {
-		EMBARC_PRINTF("Error: Wifi is not ready for lwM2M client.\r\n");
-		goto error_exit;
-	}
-
-	/* exit lwm2mClient */
-	if (c_quit == 1) {
-		EMBARC_PRINTF("Error: Try to exit existing client.\r\n");
-		handle_sigint(2);
-		goto error_exit;
-	}
-	
-	if (lwm2m_client_conn_stat == 1) {
-		EMBARC_PRINTF("Error: LwM2M client is already running.\r\n");
-		goto error_exit;
-	}
-
-	if (p_server == NULL) {
-		EMBARC_PRINTF("Error: Server Url is not specified, please check it.\r\n");
-		goto error_exit;
-	}
-
-	c_info.server = p_server;
-	c_info.ep_name = p_client_name;
-	c_info.serverPort = p_port;
-
-	lwm2m_client_start_flag = 1;
-	EMBARC_PRINTF("Start lwm2m client.\r\n"); 
-
-	/* create or resume task for lwm2mClient to realize communication with iBaby Smarthome Gateway */
-	if (xTaskCreate(task_lwm2m_client, "lwm2m client", STACK_DEPTH_LWM2M, NULL, TSKPRI_HIGH, 
-		&task_lwm2m_client_handle) != pdPASS) {
-		EMBARC_PRINTF("Error: Create task_lwm2m_client failed\r\n");
-		return E_SYS;
-	} else {
-		vTaskResume(task_lwm2m_client_handle);
-		return E_OK;
-	}
-
-	return E_OK;
-
-	error_exit:
-	return E_SYS;
-}
-#endif/* LWM2M_CLIENT */
 
 /** @} */
