+++ ibaby_smarthome_multinode/src/wearable_node/function/process_hrate.h
@@ -1,0 +1,6 @@
+#ifndef PROCESS_HRATE_H
+#define PROCESS_HRATE_H
+
+extern void process_hrate(uint32_t *hrate);
+
+#endif
