+++ ibaby_smarthome_multinode/src/wearable_node/lwm2m_client/object_motion_intensity.c
@@ -78,10 +78,11 @@
 #include <ctype.h>
 #include "dev_iic.h"
 #include "board.h"
-#include "value.h"
 
 #include "embARC.h"
 #include "embARC_debug.h"
+
+#include "common.h"
 
 
 #define PRV_RESOURCE_3_SIZE 190
